Move test utility modules into `test/utils/`
This patch moves all test utility modules into `test/utils/` and the
tests for these modules into `test/utils/test/`.

This is to make test utilities more organized and in preperation for
adding more test utilities related to test suites.
