+++ test/test_parsers/test_nquads_w3c.py
@@ -4,7 +4,7 @@ test suite."""
 
 import os
 from test.data import TEST_DATA_DIR
-from test.manifest import RDFT, RDFTest, read_manifest
+from test.utils.manifest import RDFT, RDFTest, read_manifest
 from typing import Callable, Dict
 
 import pytest
