+++ test/test_parsers/test_nt_w3c.py
@@ -3,7 +3,7 @@ test suite."""
 import os
 from test import TEST_DIR
 from test.data import TEST_DATA_DIR
-from test.manifest import RDFT, RDFTest, read_manifest
+from test.utils.manifest import RDFT, RDFTest, read_manifest
 from typing import Callable, Dict
 
 import pytest
