+++ test/test_store/test_namespace_binding.py
@@ -3,7 +3,7 @@ import itertools
 import logging
 from dataclasses import dataclass, field
 from pathlib import Path
-from test.testutils import pytest_mark_filter
+from test.utils import pytest_mark_filter
 from typing import Any, Callable, Dict, Set, Union
 
 import pytest
