+++ test/utils/test/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains tests for test utility modules inside `test.utils` (i.e. ``test/utils/``).
+"""
